Code Review Changes
Changes from the code review
