+++ src/devices/LiquidLevel/Llc200d3sh.cs
@@ -4,21 +4,21 @@
 
 using System.Device.Gpio;
 
-namespace Iot.Device.LiquidLevelSwitch
+namespace Iot.Device.LiquidLevel
 {
     /// <summary>
     /// Optomax LLC200D3SH digital liquid level switch
     /// </summary>
-    public class Llc200d3sh : LiquidLevelSwitchBase
+    public class Llc200d3sh : LiquidLevelSwitch
     {
-        /// <summary>Initializes a new instance of the <see cref="Llc200d3sh" /> class.</summary>
+        /// <summary>Creates a new instance of the Llc200d3sh.</summary>
         /// <param name="pin">The data pin</param>
         /// <param name="pinNumberingScheme">Use the logical or physical pin layout</param>
         /// <param name="gpioController">A Gpio Controller if you want to use a specific one</param>
         /// <param name="shouldDispose">True to dispose the Gpio Controller</param>
         public Llc200d3sh(int pin, GpioController gpioController = null,
                           PinNumberingScheme pinNumberingScheme = PinNumberingScheme.Logical, bool shouldDispose = true)
-            : base(pin, PinMode.Low, gpioController, pinNumberingScheme, shouldDispose)
+            : base(pin, PinValue.Low, gpioController, pinNumberingScheme, shouldDispose)
         {
         }
     }
