+++ src/devices/LiquidLevel/samples/LiquidLevelSwitch.Sample.csproj
@@ -1,0 +1,23 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>netcoreapp2.1</TargetFramework>
+    <EnableDefaultItems>false</EnableDefaultItems>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <None Remove="rpi-llc200d3sh.fzz" />
+    <None Remove="rpi-llc200d3sh_bb.png" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Compile Include="LiquidLevelSwitch.sample.cs" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\LiquidLevel.csproj" />
+    <ProjectReference Include="$(MainLibraryPath)System.Device.Gpio.csproj" />
+  </ItemGroup>
+  
+</Project>
