+++ src/devices/LiquidLevel/samples/LiquidLevelSwitch.sample.cs
@@ -3,15 +3,17 @@
 // See the LICENSE file in the project root for more information.
 
 using System;
+using System.Device.Gpio;
 using System.Threading;
+using Iot.Device.LiquidLevel;
 
-namespace Iot.Device.LiquidLevelSwitch.Samples
+namespace LiquidLevel.Sample
 {
     internal class Program
     {
         public static void Main(string[] args)
         {
-            using (Llc200d3sh sensor = new Llc200d3sh(23))
+            using (LiquidLevelSwitch sensor = new LiquidLevelSwitch(23, PinValue.Low))
             {
                 while (true)
                 {
