--- src/devices/LiquidLevelSwitch/README.md
@@ -1,17 +1,0 @@
-# Digital liquid level switch
-
-Digital liquid level switches are devices that can detect the presence of liquid/water. GPIO can be used to communicate with the devices.
-
-## Devices
-
-The implementation supports the following devices:
-
-- LLC200D3SH sensor ([Datasheet](https://cdn-shop.adafruit.com/product-files/3397/3397_datasheet_actual.pdf))
-
-## Usage
-
-An example on how to use this device binding is available in the [samples](samples) folder.
-
-The following fritzing diagram illustrates one way to wire up the Optomax LLC200D3SH digital liquid level switch with a Raspberry Pi.
-
-![Raspberry Pi Breadboard diagram](samples/rpi-llc200d3sh_bb.png)
